Fix document editor cache issue. The editors get the value from the `useState` inside `useDocumentValueContext`. This state is set from the props comming from the backend. Somehow editing a document, going other page and comming back is showing stale text in the text editor. Now we get that info fresh from the `swr` store for the old and new editors
