+++ apps/web/src/app/(private)/projects/[projectId]/versions/[commitUuid]/documents/[documentUuid]/_components/DocumentEditor/Editor/OldDocumentEditor.tsx
@@ -8,7 +8,6 @@ import {
 } from '$/hooks/useDocumentValueContext'
 import { useIsLatitudeProvider } from '$/hooks/useIsLatitudeProvider'
 import { useMetadata } from '$/hooks/useMetadata'
-import useDocumentVersions from '$/stores/documentVersions'
 import useProviderApiKeys from '$/stores/providerApiKeys'
 import { DocumentVersion, ProviderApiKey } from '@latitude-data/core/browser'
 import { SplitPane } from '@latitude-data/web-ui/atoms/SplitPane'
@@ -16,7 +15,6 @@ import {
   useCurrentCommit,
   useCurrentProject,
 } from '@latitude-data/web-ui/providers'
-import { useMemo } from 'react'
 import { EvaluationEditorHeader } from '../../../(withTabs)/evaluations/[evaluationUuid]/editor/_components/EvaluationEditor/EditorHeader'
 import DocumentTabs from '../../DocumentTabs'
 import { PlaygroundBlocksEditor } from './BlocksEditor'
@@ -40,7 +38,10 @@ export function OldDocumentEditor(props: DocumentEditorProps) {
   return (
     <MetadataProvider>
       <DevModeProvider>
-        <DocumentValueProvider document={props.document}>
+        <DocumentValueProvider
+          document={props.document}
+          documents={props.documents}
+        >
           <OldDocumentEditorContent {...props} />
         </DocumentValueProvider>
       </DevModeProvider>
@@ -49,8 +50,6 @@ export function OldDocumentEditor(props: DocumentEditorProps) {
 }
 
 function OldDocumentEditorContent({
-  document: _document,
-  documents: _documents,
   providerApiKeys,
   freeRunsCount,
   copilotEnabled,
@@ -64,22 +63,8 @@ function OldDocumentEditorContent({
   const { data: providers } = useProviderApiKeys({
     fallbackData: providerApiKeys,
   })
-  const { data: documents } = useDocumentVersions(
-    {
-      commitUuid: commit.uuid,
-      projectId: project.id,
-    },
-    {
-      fallbackData: _documents,
-    },
-  )
-
-  const document = useMemo(
-    () =>
-      documents?.find((d) => d.documentUuid === _document.documentUuid) ??
-      _document,
-    [documents, _document],
-  )
+  const { document, value, setValue, updateDocumentContent, isSaved } =
+    useDocumentValue()
   const oldHeaderEditorActions = useOldEditorHeaderActions({
     project: useCurrentProject().project,
     commit: useCurrentCommit().commit,
@@ -93,7 +78,6 @@ function OldDocumentEditorContent({
   })
   const isMerged = commit.mergedAt !== null
   const { devMode } = useDevMode()
-  const { value, setValue, updateDocumentContent, isSaved } = useDocumentValue()
   const { customReadOnlyMessage, highlightedCursorIndex } = useLatteStreaming({
     value,
     setValue,
