+++ apps/web/src/app/api/latte/debug/versions/route.ts
@@ -15,7 +15,7 @@ export const GET = errorHandler(
       },
     ) => {
       const result = await getLatteDebugVersions(workspace.id)
-      const latteVersions = result.unwrap()
+      const latteVersions = result.error ? [] : result.unwrap()
 
       return NextResponse.json(latteVersions, { status: 200 })
     },
