+++ apps/web/src/hooks/useDocumentValueContext.tsx
@@ -22,6 +22,7 @@ import {
   useCallback,
   useContext,
   useEffect,
+  useMemo,
   useState,
 } from 'react'
 import { useDebouncedCallback } from 'use-debounce'
@@ -36,6 +37,7 @@ type DocumentValueContextType = {
   value: string
   setValue: updateContentFn
   updateDocumentContent: updateContentFn
+  document: DocumentVersion
   isSaved: boolean
 }
 
@@ -46,21 +48,35 @@ const DocumentValueContext = createContext<
 type DocumentValueProviderProps = {
   children: ReactNode
   document: DocumentVersion
+  documents?: DocumentVersion[]
 }
 
 export function DocumentValueProvider({
   children,
-  document,
+  document: _document,
+  documents: _documents,
 }: DocumentValueProviderProps) {
-  const [value, setValue] = useState(document.content)
-  console.log('🐛 DocumentValueProvider mounted with:', {
-    documentUuid: document.documentUuid,
-    documentContent: document.content,
-    initialValue: value,
-    timestamp: new Date().toISOString(),
-  })
-  const { project } = useCurrentProject()
   const { commit } = useCurrentCommit()
+  const { project } = useCurrentProject()
+  const { data: documents } = useDocumentVersions(
+    {
+      commitUuid: commit.uuid,
+      projectId: project.id,
+    },
+    {
+      fallbackData: _documents,
+      keepPreviousData: true,
+      revalidateOnMount: true,
+    },
+  )
+  const document = useMemo(
+    () =>
+      documents?.find((d) => d.documentUuid === _document.documentUuid) ??
+      _document,
+    [documents, _document],
+  )
+
+  const [value, setValue] = useState(document.content)
   const { updateContent, isUpdatingContent } = useDocumentVersions({
     commitUuid: commit.uuid,
     projectId: project.id,
@@ -77,9 +93,7 @@ export function DocumentValueProvider({
   )
   const updateDocumentContent = useDebouncedCallback(
     async (content: string, opts?: Parameters<updateContentFn>[1]) => {
-      console.log('📝 Updating document content', content, opts)
       setContentValue(content, opts)
-      console.log('✅ Document content updated', content)
 
       const [_, error] = await updateContent({
         commitUuid: commit.uuid,
@@ -95,7 +109,6 @@ export function DocumentValueProvider({
           description: 'There was an error saving the document.',
           variant: 'destructive',
         })
-        console.error('🛠️ Error saving document:', error)
         setContentValue(document.content)
       }
     },
@@ -119,6 +132,7 @@ export function DocumentValueProvider({
         setValue: setContentValue,
         updateDocumentContent,
         isSaved: !isUpdatingContent,
+        document,
       }}
     >
       {children}
