+++ packages/core/src/services/copilot/latte/debugVersions.ts
@@ -1,9 +1,9 @@
-import { UnauthorizedError } from '@latitude-data/constants/errors'
+import { BadRequestError } from '@latitude-data/constants/errors'
 import { Result } from '../../../lib/Result'
 import { PromisedResult } from '../../../lib/Transaction'
+import { CommitsRepository } from '../../../repositories'
 import { isFeatureEnabledByName } from '../../workspaceFeatures/isFeatureEnabledByName'
 import { assertCopilotIsSupported, getCopilotDocument } from './helpers'
-import { CommitsRepository } from '../../../repositories'
 
 export type LatteVersion = {
   uuid: string
@@ -19,15 +19,11 @@ async function assertLatteDebugModeIsEnabled(
     'latteDebugMode',
   )
 
-  // TODO: This returns `boolean | Result<boolean>` for some reason
-  const isEnabled =
-    typeof isEnabledResult === 'boolean'
-      ? isEnabledResult
-      : isEnabledResult.value
+  if (!Result.isOk(isEnabledResult)) return isEnabledResult
 
-  if (!isEnabled) {
+  if (!isEnabledResult.unwrap()) {
     return Result.error(
-      new UnauthorizedError('This workspace cannot use Latte Debug Mode'),
+      new BadRequestError('This workspace cannot use Latte Debug Mode'),
     )
   }
